86cv684zq: Removed testing from homepage, added messages inbox to messages page with dummy data, adjusted padding for message-left css class
