+++ behind-the-veil-siteroot/.meteor/packages
@@ -17,7 +17,6 @@ typescript@4.9.5              # Enable TypeScript syntax in .ts and .tsx modules
 shell-server@0.5.0            # Server-side component of the `meteor shell` command
 hot-module-replacement@0.5.3  # Update client in development without reloading the page
 
-autopublish@1.0.7             # Publish all data to the clients (for prototyping)
 insecure@1.0.7                # Allow all DB writes from clients (for prototyping)
 static-html@1.3.2             # Define static page content in .html files
 react-meteor-data       # React higher-order component for reactively tracking Meteor data
