+++ behind-the-veil-siteroot/.meteor/versions
@@ -1,5 +1,4 @@
 allow-deny@1.1.1
-autopublish@1.0.7
 autoupdate@1.8.0
 babel-compiler@7.10.5
 babel-runtime@1.5.1
