+++ behind-the-veil-siteroot/imports/ui/components/pages/home/HomePage.jsx
@@ -7,8 +7,6 @@
 import React from 'react';
 import WhiteBackground from "../../whiteBackground/WhiteBackground.jsx";
 import PageLayout from "../../../enums/PageLayout";
-import MessagesPreview from '../messages/MessagesPreview.jsx';
-import MessagesInbox from '../messages/MessagesInbox.jsx';
 
 /**
  * Home page which is also the landing page
@@ -17,7 +15,6 @@ export const HomePage = () => {
     return (
         <WhiteBackground pageLayout={PageLayout.LARGE_CENTER}>
             <span>Home Page to be done!!</span>
-            <MessagesInbox></MessagesInbox>
         </WhiteBackground>
     );
 };
